docker/files/api/requirements.txt
appdirs==1.4.3
asn1crypto==0.22.0
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.3
click==6.7
cryptography==2.5
Flask==0.12.3
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
packaging==16.8
pycparser==2.17
PyJWT==1.5.0
pyparsing==2.2.0
requests==2.20.0
simplejson==3.10.0
six==1.10.0
urllib3==1.24.1
Werkzeug==0.12.2